// model/userModel.ts
// export interface UserPostRequest {
//     UID:      number;
//     name:     string;
//     email:    string;
//     password: string;
//     image:    string;
//     type:     string;
// }

export interface UserPostRequest {
    UID:      number;
    name:     string;
    email:    string;
    password: string;
    image:    string;
    type:     string;
}